requirements.txt
pytube>=15.0.0
youtube-transcript-api>=0.6.1
streamlit>=1.32.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0